# Optimize BigInteger.CompareTo()