--- a/src/libraries/System.Runtime.Numerics/src/System/Numerics/BigIntegerCalculator.Utils.cs
+++ b/src/libraries/System.Runtime.Numerics/src/System/Numerics/BigIntegerCalculator.Utils.cs
@@ -1,6 +1,10 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
+using System.Diagnostics;
+using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
+
 namespace System.Numerics
 {
     internal static partial class BigIntegerCalculator
@@ -15,19 +19,20 @@ namespace System.Numerics
 
         public static int Compare(ReadOnlySpan<uint> left, ReadOnlySpan<uint> right)
         {
-            if (left.Length < right.Length)
-                return -1;
-            if (left.Length > right.Length)
-                return 1;
+            Debug.Assert(left.Length == 0 || left[^1] > 0);
+            Debug.Assert(right.Length == 0 || right[^1] > 0);
+
+            if (left.Length != right.Length)
+                return left.Length < right.Length ? -1 : 1;
+
+            ref uint rightPtr = ref MemoryMarshal.GetReference(right);
 
             for (int i = left.Length - 1; i >= 0; i--)
             {
                 uint leftElement = left[i];
-                uint rightElement = right[i];
-                if (leftElement < rightElement)
-                    return -1;
-                if (leftElement > rightElement)
-                    return 1;
+                uint rightElement = Unsafe.Add(ref rightPtr, i);
+                if (leftElement != rightElement)
+                    return leftElement < rightElement ? -1 : 1;
             }
 
             return 0;